requirements.txt
streamlit
pandas
openpyxl
pdfplumber
pytesseract
Pillow
opencv-python-headless
spacy
numpy
